Add files via upload
These codes are used to generate the Lip_Bengal Dataset from raw videos

Where :

i. folders are made according to a text file

ii. extracted image frames are generated from videos according to annotation done in xl file

iii. And those frames are saved to desired destination folders.

To run the code as it is just put raw videos in the raw data folder and change the directories in the notebook and you will be good to go.